+++ update-api/app/Core/Router.php
@@ -96,6 +96,12 @@ class Router
                 } elseif (is_callable($routeInfo[1])) {
                     call_user_func($routeInfo[1]);
                 }
+                if ($route !== '/login' && !$isApi) {
+                    if (!SessionManager::getInstance()->requireAuth()) {
+                        return;
+                    }
+                }
+                call_user_func_array([new $class(), $action], $vars);
                 break;
         }
     }